requirements.txt
Django==2.2.7
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
et-xmlfile==1.1.0
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
gunicorn==20.1.0
psycopg2>=2.8,<2.9
